+++ NEWS.md
@@ -7,6 +7,7 @@
 BUGFIXES :    
 
 * `.st_mandatory_params()` property efficiencywithdrawal is not a ratio and can be greater than 1.
+* `createStudy()` Since Antares 9.2, property `initial-reservoir-levels` in `generaldata/other preferences` is not needed. This property must be removed.
 
 
 # antaresEditObject 1.0.0