+++ NEWS.md
@@ -8,6 +8,11 @@ ENHANCEMENT :
 
 * `importZipStudyWeb()` zip study with new package `zip`instead of `utils::zip` (more faster)
 
+BUGFIXES :    
+
+* `.st_mandatory_params()` property efficiencywithdrawal is not a ratio and can be greater than 1.
+
+
 # antaresEditObject 1.0.0
 
 NEW FEATURES :